third/centralblocklist v1.0.0: For users upgrading: requires config changes!
The api-key is now no longer set in this module. It requires another module
which is used for both acquiring and setting api keys.
See https://www.unrealircd.org/docs/Central_API

Other changes:
* API version bumped to v1
* CBL tests are now bundles every second. So if 10 users connect in 1 second
  we will only bother the CBL server once. And it will return results for
  all 10 users in 1 response.
* CBL can now return "errors"
* CBL can now return "warnings"
* CBL can now set tags on users, not only CBL_SCORE but also other ones
  in the future. These can then be used by for example central spamfilter
  or at other places that use ::rule's.

All this is work in progress. Now our test users can try this v1.0.0,
which will likely receive bug fixes as a result. And then hopefully we
can open to the public at a later point.