Add enhanced SPAMREPORT command, to report spam via Central API. This is work in progress, the server-side is not done yet.
To enable this setting is required:
set {
	central-blocklist {
		spam-report yes;
	}
}

Then, as an IRCOp, type /SPAMREPORT <nick> this will send a spam report
about the user to us. This includes the handshake data, which the
centralblocklist module already provides, but also includes the last
10 commands executed. Only PRIVMSG/NOTICE/PART/NICK commands are sent,
and for PRIVMSG/NOTICE only messages to channels not private messages.

This commit also fixes two memory leaks in centralblocklist submissions.