auditorium: add compatibility with Unreal >= 6.2.1, also remove the isOper() exception for CAN_SEND_TO_CHANNEL because I think that was never correct in the first place :D
Note that specifically on Unreal 6.2.1, changing a channel mode while +u is in effect may cause +d to be set. This should be fixed starting from Unreal 6.2.2. The isOper() exception caused regular users to see messages from opers without +oaq, but the oper still wasn't visible in the channel. That's pretty weird and actually doesn't even match with the module description.